Changes to be committed:

+++ smarts/core/sumo_traffic_simulation.py
@@ -90,8 +90,6 @@ class SumoTrafficSimulation:
         self._reserved_areas = dict()
         self._allow_reload = allow_reload
 
-        atexit.register(self._destroy)
-
     def __repr__(self):
         return f"""SumoTrafficSim(
   _scenario={repr(self._scenario)},
@@ -109,11 +107,10 @@ class SumoTrafficSimulation:
     def __str__(self):
         return repr(self)
 
-    def _destroy(self):
-        if atexit:
-            atexit.unregister(self._destroy)
-
+    def destroy(self):
         self._close_traci_and_pipes()
+        if not self._is_setup:
+            return
         self._sumo_proc.terminate()
         self._sumo_proc.wait()
